+++ examples/nextjs/pages/_app.tsx
@@ -1,5 +1,5 @@
 import '../styles/globals.css';
-import type { AppProps } from 'next/app';
+import type { AppProps } from 'next/app.js';
 import '@livekit/components-styles';
 
 function MyApp({ Component, pageProps }: AppProps) {