+++ examples/nextjs/pages/e2ee.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { LiveKitRoom, useToken, VideoConference, setLogLevel } from '@livekit/components-react';
 import type { NextPage } from 'next';
 import * as React from 'react';