+++ packages/react/src/components/participant/animationSequences/listeningSequence.ts
@@ -1,0 +1,6 @@
+export const generateListeningSequenceBar = (columns: number): number[][] => {
+  const center = Math.floor(columns / 2);
+  const noIndex = -1;
+
+  return [[center], [noIndex]];
+};