+++ packages/react/src/context/index.ts
@@ -23,5 +23,4 @@ export {
   useTrackRefContext,
 } from './track-reference-context';
 
-export { FeatureFlags, useFeatureContext, LKFeatureContext } from './feature-context';
-export { VoiceAssistantContext } from './voice-assistant-context';
+export { type FeatureFlags, useFeatureContext, LKFeatureContext } from './feature-context';