+++ packages/react/src/hooks/index.ts
@@ -1,45 +1,45 @@
 export { useAudioPlayback } from './useAudioPlayback';
 export { useClearPinButton } from './useClearPinButton';
 export {
-  ConnectionQualityIndicatorOptions,
+  type ConnectionQualityIndicatorOptions,
   useConnectionQualityIndicator,
 } from './useConnectionQualityIndicator';
 export { useConnectionState } from './useConnectionStatus';
 export { useDataChannel } from './useDataChannel';
 export { useDisconnectButton } from './useDisconnectButton';
 export { useFacingMode } from './useFacingMode';
-export { UseFocusToggleProps, useFocusToggle } from './useFocusToggle';
+export { type UseFocusToggleProps, useFocusToggle } from './useFocusToggle';
 export { useGridLayout } from './useGridLayout';
-export { UseIsMutedOptions, useIsMuted } from './useIsMuted';
+export { type UseIsMutedOptions, useIsMuted } from './useIsMuted';
 export { useIsSpeaking } from './useIsSpeaking';
 export { useLiveKitRoom } from './useLiveKitRoom';
-export { UseLocalParticipantOptions, useLocalParticipant } from './useLocalParticipant';
+export { type UseLocalParticipantOptions, useLocalParticipant } from './useLocalParticipant';
 export { useLocalParticipantPermissions } from './useLocalParticipantPermissions';
-export { UseMediaDeviceSelectProps, useMediaDeviceSelect } from './useMediaDeviceSelect';
+export { type UseMediaDeviceSelectProps, useMediaDeviceSelect } from './useMediaDeviceSelect';
 export { useMediaDevices } from './useMediaDevices';
 export { usePagination } from './usePagination';
-export { UseParticipantInfoOptions, useParticipantInfo } from './useParticipantInfo';
+export { type UseParticipantInfoOptions, useParticipantInfo } from './useParticipantInfo';
 export {
-  UseParticipantPermissionsOptions,
+  type UseParticipantPermissionsOptions,
   useParticipantPermissions,
 } from './useParticipantPermissions';
-export { UseParticipantTileProps, useParticipantTile } from './useParticipantTile';
-export { UseParticipantsOptions, useParticipants } from './useParticipants';
+export { type UseParticipantTileProps, useParticipantTile } from './useParticipantTile';
+export { type UseParticipantsOptions, useParticipants } from './useParticipants';
 export { usePinnedTracks } from './usePinnedTracks';
-export { UseRemoteParticipantOptions, useRemoteParticipant } from './useRemoteParticipant';
-export { UseRemoteParticipantsOptions, useRemoteParticipants } from './useRemoteParticipants';
-export { UseRoomInfoOptions, useRoomInfo } from './useRoomInfo';
+export { type UseRemoteParticipantOptions, useRemoteParticipant } from './useRemoteParticipant';
+export { type UseRemoteParticipantsOptions, useRemoteParticipants } from './useRemoteParticipants';
+export { type UseRoomInfoOptions, useRoomInfo } from './useRoomInfo';
 export { useSortedParticipants } from './useSortedParticipants';
 export { useSpeakingParticipants } from './useSpeakingParticipants';
-export { UseStartAudioProps, useStartAudio } from './useStartAudio';
-export { UseStartVideoProps, useStartVideo } from './useStartVideo';
-export { UseSwipeOptions, useSwipe } from './useSwipe';
-export { UseChatToggleProps, useChatToggle } from './useChatToggle';
-export { UseTokenOptions, UserInfo, useToken } from './useToken';
+export { type UseStartAudioProps, useStartAudio } from './useStartAudio';
+export { type UseStartVideoProps, useStartVideo } from './useStartVideo';
+export { type UseSwipeOptions, useSwipe } from './useSwipe';
+export { type UseChatToggleProps, useChatToggle } from './useChatToggle';
+export { type UseTokenOptions, type UserInfo, useToken } from './useToken';
 export { useTrackMutedIndicator } from './useTrackMutedIndicator';
-export { UseTrackToggleProps, useTrackToggle } from './useTrackToggle';
-export { UseTracksHookReturnType, UseTracksOptions, useTracks } from './useTracks';
-export { UseVisualStableUpdateOptions, useVisualStableUpdate } from './useVisualStableUpdate';
+export { type UseTrackToggleProps, useTrackToggle } from './useTrackToggle';
+export { type UseTracksHookReturnType, type UseTracksOptions, useTracks } from './useTracks';
+export { type UseVisualStableUpdateOptions, useVisualStableUpdate } from './useVisualStableUpdate';
 // export { UseTrackOptions, useTrack } from './useTrack';
 export { useTrackByName } from './useTrackByName';
 export { useChat } from './useChat';
@@ -47,7 +47,7 @@ export {
   usePersistentUserChoices,
   type UsePersistentUserChoicesOptions,
 } from './usePersistentUserChoices';
-export { UseIsEncryptedOptions, useIsEncrypted } from './useIsEncrypted';
+export { type UseIsEncryptedOptions, useIsEncrypted } from './useIsEncrypted';
 export * from './useTrackVolume';
 export * from './useParticipantTracks';
 export * from './useTrackTranscription';